refactor(workspace): extract quspin-matrix + quspin-expm (PR 5+6/7)
Combines spec PRs 5 and 6 because Hamiltonian::expm_dot inside
quspin-matrix requires expm_multiply_auto at the same crate level —
extracting them together avoids a temporary quspin-core cycle.

## quspin-matrix

Absorbs:
- qmatrix/ module (matrix, build, dispatch, ops)
- hamiltonian/ module (ham, inner, schrodinger)
- operator/apply.rs → apply.rs (generic apply_and_project_to)
- linear_operator/qmatrix_op.rs → qmatrix_op.rs (LinearOperator adapter)
- OperatorDispatch trait + impls for *OperatorInner → dispatch.rs
  (previously impl'd per-enum in quspin-core/src/operator/*/dispatch.rs)
- trace/one-norm helpers → qmatrix_helpers.rs (pub(crate))

quspin-matrix depends on quspin-types, quspin-bitbasis, quspin-operator,
quspin-basis, quspin-expm. It is the "glue" crate — the only one that
sees both operators and bases.

## quspin-expm

Absorbs expm/{algorithm,norm_est,params}.rs (minus compute.rs which
stays in quspin-types and linear_operator.rs shim which is deleted).
Rewrites mod.rs as a thin lib.rs that exposes:

    pub fn expm_multiply_auto<V, Op>(...)   where Op: LinearOperator<V>
    pub fn expm_multiply_auto_into<V, Op>(...)
    pub fn expm_multiply_many_auto<V, Op>(...)
    pub fn expm_multiply_many_auto_into<V, Op>(...)
    pub fn expm_multiply<V, Op>(...)        (low-level, explicit params)
    pub fn expm_multiply_many<V, Op>(...)
    pub fn expm_multiply_par<V, Op>(...)

quspin-expm depends only on quspin-types — it is oblivious to QMatrix,
Hamiltonian, and basis types, satisfying the spec's §3.1 requirement
that expm compile in parallel with quspin-matrix.

## Deleted from quspin-core

All qmatrix/, hamiltonian/ sources plus expm/{algorithm,linear_operator,
norm_est,params}.rs and linear_operator/qmatrix_op.rs. The remaining
files under those dirs are pub-use shims.

## Test impact

- cargo test --workspace: 279 tests pass
- Python tests (113 fast): all pass
- Ten Rust integration tests in the old expm/mod.rs were tests of
  QMatrixOperator + expm_multiply_auto composition; they belong in
  quspin-matrix and will be reinstated in a follow-up PR.
