+++ Cargo.toml
@@ -4,6 +4,8 @@ members = [
     "crates/quspin-bitbasis",
     "crates/quspin-operator",
     "crates/quspin-basis",
+    "crates/quspin-matrix",
+    "crates/quspin-expm",
     "crates/quspin-core",
     "crates/quspin-py",
 ]
