+++ crates/quspin-core/src/hamiltonian/mod.rs
@@ -1,7 +1,2 @@
-pub mod ham;
-pub mod inner;
-pub mod schrodinger;
-
-pub use ham::{CoeffFn, Hamiltonian};
-pub use inner::{HamiltonianInner, IntoHamiltonianInner};
-pub use schrodinger::SchrodingerEq;
+// Forwarding shim — hamiltonian types now live in `quspin-matrix`.
+pub use quspin_matrix::hamiltonian::*;
