+++ crates/quspin-core/src/operator/fermion/mod.rs
@@ -1,5 +1,3 @@
-pub mod dispatch;
-
-pub use dispatch::FermionOperatorInner;
-pub use quspin_operator::fermion::operator;
-pub use quspin_operator::fermion::{FermionOp, FermionOpEntry, FermionOperator};
+pub use quspin_operator::fermion::{
+    FermionOp, FermionOpEntry, FermionOperator, FermionOperatorInner, operator,
+};
