+++ crates/quspin-core/src/operator/spin/mod.rs
@@ -1,5 +1,1 @@
-pub mod dispatch;
-
-pub use dispatch::SpinOperatorInner;
-pub use quspin_operator::spin::operator;
-pub use quspin_operator::spin::{SpinOp, SpinOpEntry, SpinOperator};
+pub use quspin_operator::spin::{SpinOp, SpinOpEntry, SpinOperator, SpinOperatorInner, operator};
