+++ crates/quspin-matrix/src/hamiltonian/mod.rs
@@ -1,0 +1,7 @@
+pub mod ham;
+pub mod inner;
+pub mod schrodinger;
+
+pub use ham::{CoeffFn, Hamiltonian};
+pub use inner::{HamiltonianInner, IntoHamiltonianInner};
+pub use schrodinger::SchrodingerEq;
